+++ core/include/interface/Interface_Transitions.h
@@ -6,7 +6,7 @@
 struct State;
 
 DLLEXPORT void Transition_Homogeneous(State *state, int idx_1, int idx_2, int idx_chain=-1);
-DLLEXPORT void Transition_Add_Noise_Temperature(State *state, double temperature, int idx_1, int idx_2, int idx_chain=-1);
+DLLEXPORT void Transition_Add_Noise_Temperature(State *state, float temperature, int idx_1, int idx_2, int idx_chain=-1);
 
 #include "DLL_Undefine_Export.h"
 #endif
